pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mp4_transcriber"
version = "0.1.0"
description = "A tool to convert MP4 videos to text transcripts using local compute"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "openai-whisper>=20230314",
    "nltk>=3.8.1",
    "numpy>=1.24.0",
]

[project.urls]
"Homepage" = "https://github.com/davehague/mp4_transcriber"
"Bug Tracker" = "https://github.com/davehague/mp4_transcriber/issues"

[project.scripts]
mp4-to-transcript = "mp4_transcriber.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
testpaths = ["tests"]
